Added more comments and Functionality
The user can now set up the start node and goal nodes. If the start node
is a wall the user can try again. If the start node has all adjacent
nodes as walls then the user will be notified and can change the start
node if the user would like.