Hopefully Got It Right Commit
After hours and a few breaks....the sun is in the center, the earth
rotates the sun and the moon rotates the earth. I had trouble figuring
out the math involving in getting the earth to rotate around the sun.
Once i figured out how to do it then i did the same math for the moon to
rotate the earth. The moon rotated correctly ONLY when the earth would
not move. When the earth would move it would fling the moon around like
crazy. I realized itw as my math that was incorrect. The final solution
was instead of taking the [3] index vector of each matrix..I simply just
set the matrix to another matrix time the translation i want it to move
by and then..poof..it worked. On a side note, It takes 3.5 sec from the
time i click for the program to start until the objects actually get
drawn to the screen so this was causing major snapping issues of the
planets. So i reduce the time variable by 3.5 seconds to accommodate for
the delay and now the planets start where they are supposed to.
I have an abstract class that is used as a base class for another class
to create the solar system and make the main function have little code
and planets rotate accordingly.