Morning assignment of 09/06/16
09/06/16
Write a function that takes a function as a parameter.The function it
receives should take two arguments(min, max).
Calculate a random number between those two numbers and return it.If the
function fails return or assert on the call.
If the function that takes the min, max function fails give feedback to
the user.